genesis-agent/agent_backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.10.3
google-cloud-aiplatform==1.116.0
google-cloud-firestore==2.19.0
google-cloud-storage==2.18.2
google-cloud-speech==2.28.0
google-cloud-texttospeech==2.18.0
google-cloud-videointelligence==2.13.5
google-cloud-vision==3.7.0
google-genai>=0.3.0
httpx==0.28.1
python-multipart==0.0.20
google-auth>=2.23.0
opencv-python>=4.8.0
moviepy>=1.0.3
pillow>=10.0.0
numpy>=1.24.0
